+++ src/Ratiocinia.Models/CheckedMultiplyPolicy.cs
@@ -11,13 +11,14 @@ namespace Ratiocinia.Models
     public sealed class CheckedMultiplyPolicy<T> :
         ICheckedDivisionFunctions<T>,
         ICheckedMultiplyFunctions<T>,
-        IEquatableGreatestCommonDivisorFunctions<T>
+        IComparableGreatestCommonDivisorFunctions<T>
         where T :
         IAdditiveIdentity<T, T>,
         IDivisionOperators<T, T, T>,
-        IEquatable<T>,
+        IComparable<T>,
         IModulusOperators<T, T, T>,
-        IMultiplyOperators<T, T, T>
+        IMultiplyOperators<T, T, T>,
+        IUnaryNegationOperators<T, T>
     {
         /// <summary>
         /// Gets the singleton instance of the <see cref="CheckedMultiplyPolicy{T}" /> class.