+++ src/Ratiocinia.Models/UncheckedMultiplyPolicy.cs
@@ -9,15 +9,16 @@ namespace Ratiocinia.Models
     /// </summary>
     /// <typeparam name="T">The numeric type for which arithmetic operations are provided.</typeparam>
     public sealed class UncheckedMultiplyPolicy<T> :
-        IEquatableGreatestCommonDivisorFunctions<T>,
+        IComparableGreatestCommonDivisorFunctions<T>,
         IUncheckedDivisionFunctions<T>,
         IUncheckedMultiplyFunctions<T>
         where T :
         IAdditiveIdentity<T, T>,
         IDivisionOperators<T, T, T>,
-        IEquatable<T>,
+        IComparable<T>,
         IModulusOperators<T, T, T>,
-        IMultiplyOperators<T, T, T>
+        IMultiplyOperators<T, T, T>,
+        IUnaryNegationOperators<T, T>
     {
         /// <summary>
         /// Gets the singleton instance of the <see cref="UncheckedMultiplyPolicy{T}" /> class.